requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Embeddings and ML
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# LLM
openai>=1.3.0

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Monitoring
prometheus-client>=0.19.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

